Fix negation of unsigned value warning treated as error
Reviewed-on: https://swiftshader-review.googlesource.com/c/SwiftShader/+/31512
Tested-by: Nicolas Capens <nicolascapens@google.com>

+++ src/Pipeline/SpirvShader.cpp
@@ -5875,7 +5875,7 @@ namespace sw
 				break;
 
 			case spv::OpSNegate:
-				v = -l;
+				v = -(int)l;
 				break;
 			case spv::OpNot:
 			case spv::OpLogicalNot:
